requirements.txt
torch
transformers
Pillow
openai
python-dotenv
streamlit